refactor(components/atom/badge): api tab
